move create profile view into views package
