+++ apple/OmnivoreKit/Sources/App/Views/RegistrationView.swift
@@ -129,7 +129,7 @@ struct RegistrationView: View {
   var body: some View {
     if let registrationState = viewModel.registrationState {
       if case let RegistrationViewModel.RegistrationState.createProfile(userProfile) = registrationState {
-        CreateProfileView(userProfile: userProfile, dataService: dataService)
+        CreateProfileContainerView(userProfile: userProfile, dataService: dataService)
       } else if case let RegistrationViewModel.RegistrationState.newAppleSignUp(userProfile) = registrationState {
         NewAppleSignupView(
           userProfile: userProfile,
